Add a message when a test was started / finished
I had the problem, that a failure in my application lead to a total
crash of the test system. Unfortunately, UT++ did not show any location
of the error or the started test. What I did was to output, e.g.:

  Executing test: [68] 'IOHELPER_TESTS'::'fullPath'
   when a test was started. This shows a number, the suite and the test id.

and

  .. Finished [68]
   when the test finished.

This way, I can at least determine the latest test in use.